Move actions to /(auth)/logout
