+++ app/(auth)/logout/actions.ts
@@ -1,6 +1,6 @@
 'use server';
 import { cookies } from 'next/headers';
-import { deleteSessionByToken } from '../database/sessions';
+import { deleteSessionByToken } from '../../../database/sessions';
 
 export async function logout() {
   const cookieStore = cookies();
